+++ src/ExchangeRateProvider/PDOProviderConfiguration.php
@@ -15,40 +15,40 @@ final class PDOProviderConfiguration
         /**
          * The name of the table that holds the exchange rates. Required.
          */
-        private string $tableName,
+        public readonly string $tableName,
 
         /**
          * The name of the column that holds the exchange rate for the currency pair. Required.
          */
-        private string $exchangeRateColumnName,
+        public readonly string $exchangeRateColumnName,
 
         /**
          * The source currency code, if it is fixed. Optional.
          *
          * If not set, $sourceCurrencyColumnName must be set.
          */
-        private ?string $sourceCurrencyCode = null,
+        public readonly ?string $sourceCurrencyCode = null,
 
         /**
          * The name of the column that holds the source currency code. Optional.
          *
          * If not set, $sourceCurrencyCode must be set.
          */
-        private ?string $sourceCurrencyColumnName = null,
+        public readonly ?string $sourceCurrencyColumnName = null,
 
         /**
          * The target currency code, if it is fixed. Optional.
          *
          * If not set, $targetCurrencyColumnName must be set.
          */
-        private ?string $targetCurrencyCode = null,
+        public readonly ?string $targetCurrencyCode = null,
 
         /**
          * The name of the column that holds the target currency code. Optional.
          *
          * If not set, $targetCurrencyCode must be set.
          */
-        private ?string $targetCurrencyColumnName = null,
+        public readonly ?string $targetCurrencyColumnName = null,
 
         /**
          * Extra WHERE conditions that will be included in the database query. Optional.
@@ -59,7 +59,7 @@ final class PDOProviderConfiguration
          *
          * This can be used, for example, to query an exchange rate for a particular date.
          */
-        private ?string $whereConditions = null,
+        public readonly ?string $whereConditions = null,
     ) {
         if ($sourceCurrencyCode === null && $sourceCurrencyColumnName === null) {
             throw new InvalidArgumentException(
@@ -90,40 +90,5 @@ final class PDOProviderConfiguration
                 'Invalid configuration: $sourceCurrencyCode and $targetCurrencyCode cannot be both set.',
             );
         }
-    }
-
-    public function getTableName(): string
-    {
-        return $this->tableName;
-    }
-
-    public function getExchangeRateColumnName(): string
-    {
-        return $this->exchangeRateColumnName;
-    }
-
-    public function getSourceCurrencyColumnName(): ?string
-    {
-        return $this->sourceCurrencyColumnName;
-    }
-
-    public function getSourceCurrencyCode(): ?string
-    {
-        return $this->sourceCurrencyCode;
-    }
-
-    public function getTargetCurrencyColumnName(): ?string
-    {
-        return $this->targetCurrencyColumnName;
-    }
-
-    public function getTargetCurrencyCode(): ?string
-    {
-        return $this->targetCurrencyCode;
-    }
-
-    public function getWhereConditions(): ?string
-    {
-        return $this->whereConditions;
     }
 }